update folder structure, and break bigger .js files down (#12)
* add to-do comment

* rename css files and place them in /assets/styles

* update folder structure with models, helpers and views

models define structure and behaviour of projects and tasks objects through factories. helpers contains the localStorage helper functions used to interact with localStorage. App.js is the central controller, managing user interactions etc (non-UI-rendering logic). Views is the view of the application. if i need to, i can always create a components/ subfolder for each of the components (eg sidebar, header, main section)

* switch to a controllers/ subfolder to hold app controller and screen controller in subfolders

might do a barrel export

* rename screen controller subfolder entry file as index.js for clarity and also reduce text when importing

* update class to ensure styling of add project modal's button is correct

* separate out addProject UI manipulation from screencontroller

* add icons again for future use. break screenController.js down into more modules

* add icons again for future use. break screenController.js down into more modules

* fix incorrect import
