+++ src/index.html
@@ -181,7 +181,7 @@
         <small class="error-message">Project name exists. Please choose a different project name.</small>
       </div>
       <div class="btn-group">
-        <button class="confirm-add-project-btn" type="submit">Add</button>
+        <button class="confirm-btn add-project-btn" type="submit">Add</button>
         <button class="cancel-btn" type="reset">Cancel</button>
       </div>
     </form>
